feat: change API interface
