+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@context-query/core",
-  "version": "0.0.7",
+  "version": "0.0.8",
   "type": "module",
   "license": "MIT",
   "scripts": {
