+++ packages/playground/package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@context-query/react": "workspace:*",
+    "@context-query/react": "latest",
     "@radix-ui/react-slot": "^1.1.2",
     "@tailwindcss/vite": "^4.0.13",
     "class-variance-authority": "^0.7.1",
