+++ packages/playground/src/CounterContextQueryProvider.tsx
@@ -3,10 +3,8 @@ import { createContextQuery } from "@context-query/react";
 export const {
   Provider: CounterQueryProvider,
   useContextQuery: useCounterQuery,
-} = createContextQuery({
-  initialState: {
-    count1: 0,
-    count2: 0,
-    count3: 0,
-  },
-});
+} = createContextQuery<{
+  count1: number;
+  count2: number;
+  count3: number;
+}>();
