+++ packages/react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@context-query/react",
-  "version": "0.0.8",
+  "version": "0.0.9",
   "license": "MIT",
   "description": "React bindings for ContextQuery",
   "author": "Minyeoung Seo <tjalsdud89@naver.com>",
